refactor: 메모 타입 전역이랑 도메인 단위로 분리
